% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSet.r
\name{DataSet}
\alias{DataSet}
\title{DataSet interface.}
\description{
This class defines a basis
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataSet$new()}}
\item \href{#method-addFromDataFrame}{\code{DataSet$addFromDataFrame()}}
\item \href{#method-clone}{\code{DataSet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSet$new(name, xVar, data = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addFromDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-addFromDataFrame}{}}}
\subsection{Method \code{addFromDataFrame()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSet$addFromDataFrame(df)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
