% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSets.r
\name{DataSets}
\alias{DataSets}
\title{DataSets class}
\description{
This class defines DataSets, a class holding one or more @DataSet@ objects.
}
\section{Super class}{
\code{\link[ranalyse:Node]{ranalyse::Node}} -> \code{DataSets}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataSets$new()}}
\item \href{#method-createCoreModelsFor}{\code{DataSets$createCoreModelsFor()}}
\item \href{#method-clone}{\code{DataSets$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ranalyse" data-topic="Node" data-id="addChild">}\href{../../ranalyse/html/Node.html#method-addChild}{\code{ranalyse::Node$addChild()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSets$new(name, datasets, desc = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createCoreModelsFor"></a>}}
\if{latex}{\out{\hypertarget{method-createCoreModelsFor}{}}}
\subsection{Method \code{createCoreModelsFor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSets$createCoreModelsFor(outcomes, fitters, formulas, adaptions, selection)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSets$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
