% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSource.r
\name{DataSource}
\alias{DataSource}
\title{DataSource interface.}
\description{
This class defines a \code{DataSource}.
}
\section{Super class}{
\code{\link[ranalyse:Node]{ranalyse::Node}} -> \code{DataSource}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataSource$new()}}
\item \href{#method-createDataSet}{\code{DataSource$createDataSet()}}
\item \href{#method-clone}{\code{DataSource$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ranalyse" data-topic="Node" data-id="addChild">}\href{../../ranalyse/html/Node.html#method-addChild}{\code{ranalyse::Node$addChild()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSource$new(xVarName, variableDesc = NULL, desc = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xVarName}}{character Name of x-column, must be inside \code{variableDesc}.}

\item{\code{variableDesc}}{sets::tuple A tuple of variable names with descriptions. E.g. \code{sets::tuple(name = "varName", desc = "varirable Description")}
param filterFun function\link{char, vector}, bool]   Takes as input two parameters (name and data). Must return a TRUE if the variable should be included. E.g. \code{function(n, dt) n != "person"}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createDataSet"></a>}}
\if{latex}{\out{\hypertarget{method-createDataSet}{}}}
\subsection{Method \code{createDataSet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSource$createDataSet(filterVarsFun = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
