% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSourceRDS.r
\name{DataSourceRDS}
\alias{DataSourceRDS}
\title{DataSourceRDS.}
\description{
This class defines a DataSource using a RDS file.
}
\section{Super classes}{
\code{\link[ranalyse:Node]{ranalyse::Node}} -> \code{\link[ranalyse:DataSource]{ranalyse::DataSource}} -> \code{DataSourceRDS}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataSourceRDS$new()}}
\item \href{#method-clone}{\code{DataSourceRDS$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ranalyse" data-topic="Node" data-id="addChild">}\href{../../ranalyse/html/Node.html#method-addChild}{\code{ranalyse::Node$addChild()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ranalyse" data-topic="DataSource" data-id="createDataSet">}\href{../../ranalyse/html/DataSource.html#method-createDataSet}{\code{ranalyse::DataSource$createDataSet()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSourceRDS$new(filepath, xVarName, variableDesc = NULL, nodeDesc = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filepath}}{string}

\item{\code{xVarName}}{string Variable name for x values. If column does not exists in data natural numbers are assumed.}

\item{\code{variableDesc}}{sets::tuple A tuple of variable names with descriptions. E.g. \code{sets::tuple(name = "varName", desc = "varirable Description")}}

\item{\code{nodeDesc}}{string Description for Node in Graph}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSourceRDS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
