% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitterGLM.r
\name{FitterGLM}
\alias{FitterGLM}
\title{FitterGLM interface.}
\description{
FitterGLM interface.

FitterGLM interface.
}
\section{Super classes}{
\code{\link[ranalyse:Node]{ranalyse::Node}} -> \code{\link[ranalyse:Fitter]{ranalyse::Fitter}} -> \code{FitterGLM}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FitterGLM$new()}}
\item \href{#method-clone}{\code{FitterGLM$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ranalyse" data-topic="Node" data-id="addChild">}\href{../../ranalyse/html/Node.html#method-addChild}{\code{ranalyse::Node$addChild()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ranalyse" data-topic="Fitter" data-id="fit">}\href{../../ranalyse/html/Fitter.html#method-fit}{\code{ranalyse::Fitter$fit()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitterGLM$new(
  family = stats::gaussian,
  na.action = "na.fail",
  akaikeFun = stats::AIC,
  desc = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FitterGLM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
