% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{getR6Class}
\alias{getR6Class}
\title{Return the classname of an R6 object}
\usage{
getR6Class(x, n.parent = 0)
}
\arguments{
\item{x}{object    R6 class object}

\item{n.parent}{numeric   Return n.parent higher superclass. Default 0.}
}
\value{
character Class name
}
\description{
Return the classname of an R6 object
}
