% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NodeProcessor.r
\name{NodeProcessor}
\alias{NodeProcessor}
\title{NodeProcessor that will be present in output tree}
\description{
This class defines a basis
}
\section{Super class}{
\code{\link[ranalyse:Node]{ranalyse::Node}} -> \code{NodeProcessor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NodeProcessor$new()}}
\item \href{#method-addProcessor}{\code{NodeProcessor$addProcessor()}}
\item \href{#method-clone}{\code{NodeProcessor$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ranalyse" data-topic="Node" data-id="addChild">}\href{../../ranalyse/html/Node.html#method-addChild}{\code{ranalyse::Node$addChild()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeProcessor$new(desc)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addProcessor"></a>}}
\if{latex}{\out{\hypertarget{method-addProcessor}{}}}
\subsection{Method \code{addProcessor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeProcessor$addProcessor(proc)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NodeProcessor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
