% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessorSum.r
\name{PreprocessorSum}
\alias{PreprocessorSum}
\title{PreprocessorSum class.}
\description{
It can be used to sum variables in order to create a new variable.
}
\section{Super classes}{
\code{\link[ranalyse:Node]{ranalyse::Node}} -> \code{\link[ranalyse:NodeProcessor]{ranalyse::NodeProcessor}} -> \code{\link[ranalyse:Preprocessor]{ranalyse::Preprocessor}} -> \code{PreprocessorSum}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PreprocessorSum$new()}}
\item \href{#method-clone}{\code{PreprocessorSum$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ranalyse" data-topic="Node" data-id="addChild">}\href{../../ranalyse/html/Node.html#method-addChild}{\code{ranalyse::Node$addChild()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ranalyse" data-topic="NodeProcessor" data-id="addProcessor">}\href{../../ranalyse/html/NodeProcessor.html#method-addProcessor}{\code{ranalyse::NodeProcessor$addProcessor()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ranalyse" data-topic="Preprocessor" data-id="preprocess">}\href{../../ranalyse/html/Preprocessor.html#method-preprocess}{\code{ranalyse::Preprocessor$preprocess()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ranalyse" data-topic="Preprocessor" data-id="then">}\href{../../ranalyse/html/Preprocessor.html#method-then}{\code{ranalyse::Preprocessor$then()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreprocessorSum$new(
  outputName,
  inputNames,
  deleteInputVars = FALSE,
  nodeDesc = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreprocessorSum$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
