% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableBoolean.r
\name{VariableBoolean}
\alias{VariableBoolean}
\title{VariableBoolean is a Variable specialised to x being a Boolean. See \code{base::factor}.}
\description{
VariableBoolean is a Variable specialised to x being a Boolean. See \code{base::factor}.

VariableBoolean is a Variable specialised to x being a Boolean. See \code{base::factor}.
}
\section{Super classes}{
\code{\link[ranalyse:Node]{ranalyse::Node}} -> \code{\link[ranalyse:Variable]{ranalyse::Variable}} -> \code{VariableBoolean}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VariableBoolean$new()}}
\item \href{#method-clone}{\code{VariableBoolean$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ranalyse" data-topic="Node" data-id="addChild">}\href{../../ranalyse/html/Node.html#method-addChild}{\code{ranalyse::Node$addChild()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableBoolean$new(name, vals)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariableBoolean$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
